requirements.txt
# requirements.txt

langgraph
langchain
langchain-openai
langchain-community
chromadb
unstructured[pdf]
pandas
scikit-learn
python-dotenv